consolidate extra fields (#5956)
