+++ src/actions/preview.js
@@ -3,10 +3,9 @@
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
 // @flow
-
-import { findBestMatchExpression, findClosestClass } from "../utils/ast";
+import { isConsole } from "../utils/preview";
+import { findBestMatchExpression } from "../utils/ast";
 import { getTokenLocation } from "../utils/editor";
-import { isReactComponent, isImmutable, isConsole } from "../utils/preview";
 import { isGeneratedId } from "devtools-source-map";
 import { PROMISE } from "./utils/middleware/promise";
 import { getExpressionFromCoords } from "../utils/editor/get-expression";
@@ -18,68 +17,16 @@ import {
   getSelectedSource,
   getSelectedFrame,
   getSymbols,
-  getCanRewind,
-  getSource
+  getCanRewind
 } from "../selectors";
 
 import { getMappedExpression } from "./expressions";
+import { getExtra } from "./pause";
 import { isEqual } from "lodash";
 
 import type { Action, ThunkArgs } from "./types";
 import type { ColumnPosition } from "../types";
 import type { AstLocation } from "../workers/parser";
-
-async function getReactProps(evaluate) {
-  const reactDisplayName = await evaluate(
-    "this.hasOwnProperty('_reactInternalFiber') ? " +
-      "this._reactInternalFiber.type.name : " +
-      "this._reactInternalInstance.getName()"
-  );
-
-  return {
-    displayName: reactDisplayName.result
-  };
-}
-
-async function getImmutableProps(expression: string, evaluate) {
-  const immutableEntries = await evaluate((exp => `${exp}.toJS()`)(expression));
-
-  const immutableType = await evaluate(
-    (exp => `${exp}.constructor.name`)(expression)
-  );
-
-  return {
-    type: immutableType.result,
-    entries: immutableEntries.result
-  };
-}
-
-async function getExtraProps(getState, expression, result, evaluate) {
-  const props = {};
-  if (isReactComponent(result)) {
-    const selectedFrame = getSelectedFrame(getState());
-    const source = getSource(getState(), selectedFrame.location.sourceId);
-    const symbols = getSymbols(getState(), source);
-
-    if (symbols && symbols.classes) {
-      const originalClass = findClosestClass(symbols, selectedFrame.location);
-
-      if (originalClass) {
-        props.react = { displayName: originalClass.name };
-      }
-    }
-
-    if (!props.react) {
-      props.react = await getReactProps(evaluate);
-    }
-  }
-
-  if (isImmutable(result)) {
-    props.immutable = await getImmutableProps(expression, evaluate);
-  }
-
-  return props;
-}
 
 function isInvalidTarget(target: HTMLElement) {
   if (!target || !target.innerText) {
@@ -103,21 +50,6 @@ function isInvalidTarget(target: HTMLElement) {
     cursorPos.top == 0;
 
   return invalidTarget || invalidToken || invalidType;
-}
-
-export function getExtra(expression: string, result: Object) {
-  return async ({ dispatch, getState, client, sourceMaps }: ThunkArgs) => {
-    const selectedFrame = getSelectedFrame(getState());
-    if (!selectedFrame) {
-      return;
-    }
-
-    const extra = await getExtraProps(getState, expression, result, expr =>
-      client.evaluateInFrame(expr, selectedFrame.id)
-    );
-
-    return extra;
-  };
 }
 
 export function updatePreview(target: HTMLElement, editor: any) {
